+++ buildSrc/CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- A `jvmName` DSL helper method.
+- A `jvmMain`, `jvmTest`, and `jvmArtifactId` DSL helper methods taking a `JavaVersion`.
+
 ### Fixed
 
 - Resolving dependency variants locally for e.g. `kotlin-test`.